refactor(layout): split data.py into domain/{config,queries}
Step 2a of the domain/ reorganization plan. The data.py file had
two jobs: pure data types (EntryConfig, EntryConfigDiff,
build_slot_unique_id, _EMPTY_SLOTS) and query helpers that walk
hass.config_entries (get_entry_config, get_managed_slots,
find_entry_for_lock_slot). Split along that fault line:

* domain/config.py: the pure type layer. No HomeAssistant
  dependency, no iteration over config entries -- just the typed
  view of a single entry's data and its diff.

* domain/queries.py: the accessor layer. get_entry_config reads
  the cached EntryConfig off runtime_data with a from_entry()
  fallback; the cross-entry queries (get_managed_slots,
  find_entry_for_lock_slot) walk hass.config_entries.

Importers that pulled multiple symbols now split into two import
lines, one per new location. This is the cost of the cleaner
boundary -- noted by the auditor and accepted.

tests/test_data.py -> tests/test_config.py to match the new home
of the types it tests.
