+++ custom_components/lock_code_manager/util.py
@@ -13,7 +13,7 @@ from homeassistant.helpers import entity_registry as er
 from homeassistant.helpers.issue_registry import IssueSeverity, async_create_issue
 
 from .const import DOMAIN
-from .data import build_slot_unique_id
+from .domain.config import build_slot_unique_id
 
 if TYPE_CHECKING:
     from .coordinator import LockUsercodeUpdateCoordinator
